requirements.txt
future~=0.18.2
pandas~=1.1.3
setuptools~=50.3.2
tabula-py~=2.2.0